+++ packages/server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @payfanout/server
 
+## 2.1.1
+
+### Patch Changes
+
+- Updated dependencies [e856737]
+  - @payfanout/core@4.3.0
+
 ## 2.1.0
 
 ### Minor Changes